requirements.txt
aniso8601==8.1.0
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
identify==1.5.13
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.8
pathspec==0.8.1
pluggy==0.13.1
pre-commit==2.9.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.4.1
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.0
Werkzeug==1.0.1
zipp==3.4.0
